isolates id prop accessors, removes from shared
