+++ lib/neo4j.rb
@@ -59,6 +59,7 @@ require 'neo4j/active_node/query/query_proxy_find_in_batches'
 require 'neo4j/active_node/query/query_proxy_eager_loading'
 require 'neo4j/active_node/query/query_proxy_link'
 require 'neo4j/active_node/labels'
+require 'neo4j/active_node/id_property/accessor'
 require 'neo4j/active_node/id_property'
 require 'neo4j/active_node/callbacks'
 require 'neo4j/active_node/initialize'
