+++ spec/unit/validation_spec.rb
@@ -44,6 +44,7 @@ describe Neo4j::ActiveNode::Validations do
         o.stub(:_persisted_obj).and_return(nil)
         o.stub(:serialized_properties).and_return({})
         o.serialized_properties
+        clazz.stub(:default_property_values).and_return({})
         clazz.stub(:cached_class?).and_return(false)
         clazz.should_receive(:neo4j_session).and_return(session)
         node.should_receive(:props).and_return(name: 'kalle2', age: '43')
